Update to allow usage on system not hosting the cert files.
if the webserver host extracted from the base_url in the config doesn't appear
to be the system running the script.  This script will pull the cert
from the external server to do the verification, and will not start
the webserver daemon on the host running the script.

This is more of a mess than it needs to be, I'll probably just add
an optional `external: true` under the webserver config and remove
all the checks.

This works, could just be simplified a bit.